Fix encoding, error messages, and confirmation dialogs on various user management pages.  Support entering of "friendly names" along with email addresses.
We nearly support Java 6 now -- our code will compile, run, and pass the DRT under both Java 5 and Java 6.  However, additions to the JDBC interfaces in Java 6 mean that, at the moment, we can't compile the exact same code under each (using Java 6 requires commenting out an import).  Other changes to support Java 6: timestamp.valueOf("2005-12-8") doesn't parse and it appears to disallow unboxing of variables that are explicitly initialized to null (even when intervening code assigns a legitimate value... very mysterious).

Make a few minor changes to support Eclipse compiler, courtesy SystemsX.

3162 : Adding user with friendly email causes problems
2808 : Build and test under Java 6

+++ caBIG/src/org/labkey/cabig/caBIGManager.java
@@ -39,6 +39,6 @@ public class caBIGManager
     {
         Long rows = Table.executeSingleton(caBIGSchema.getInstance().getSchema(), "SELECT COUNT(*) FROM " + caBIGSchema.getInstance().getTableInfoContainers() + " WHERE EntityId = ?", new Object[]{c.getId()}, Long.class);
 
-        return (1 == rows);
+        return (1 == rows.longValue());
     }
 }
